+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2024-??-??
 
+## [0.10.0] - 2024-07-10
+
 ### Added
 
 - `layout_with_weighted_pmds` : Compute a graph layout using weighted PMDS.